[WIP] homework - listing 44
The initial implementation assumed all we would ever do is disassemble,
the assumption being that when we implement execute we will stop
disassembling... but that's lame!

So now there needs to be an explicit decode step that stores the
fields d, w, reg, mod, rm, and the bytes... so that we can then
pass that information into disassemble... or execute... or both!