JSON parse almost working
Things are probably fine for my purposes but I started on the wrong foot
with parsing strings, treating " as a token by itself when it should really
be part of the string... and I'm parsing through the file with just
one pointer instead of a pair so I can't "put back" the string once
I've parsed it... I can rewind single character tokens but not multi.

That being said, it does what it needs to do for the project so maybe I
will take a break, finish the homework, and then come back to this later.