Switched to using attribute cleanup
I realized that I would not be able to keep track of a stack of
parents without allocating memory... unless I could piggyback
on the call stack! So that's why I did. At the start of a stack
frame I record the hashed indices of the current and previous
functions... and then I restore from this on cleanup.